Update views.py
Changing the webhook view to validate the webhook message by merely verifying that a stripe event exists with the same stripe id that was contained in the webhook message. If the stripe id exists, then the webhook message is assumed to be valid. 

This commit should solve issue #121: Events being marked as Invalid, resulting in no processing of events 

Note: The model method Event.validate can therefore be deleted, since it is no longer used.